requirements.txt
adal==1.2.7
alembic==1.4.3
anyio==3.3.4
apache-airflow==2.0.0
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==1.1.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-microsoft-azure==2.0.0
apache-airflow-providers-sqlite==2.0.1
apispec==3.3.2
argcomplete==1.12.2
attrs==20.3.0
azure-batch==11.0.0
azure-common==1.1.27
azure-core==1.20.1
azure-cosmos==3.2.0
azure-datalake-store==0.0.52
azure-identity==1.7.1
azure-keyvault==4.1.0
azure-keyvault-certificates==4.3.0
azure-keyvault-keys==4.4.0
azure-keyvault-secrets==4.3.0
azure-kusto-data==0.0.45
azure-mgmt-containerinstance==1.5.0
azure-mgmt-core==1.3.0
azure-mgmt-datafactory==1.1.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-resource==20.0.0
azure-nspkg==3.0.2
azure-storage-blob==12.8.1
azure-storage-common==2.1.0
azure-storage-file==2.1.0
Babel==2.9.0
blinker==1.4
cached-property==1.5.2
cattrs==1.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
charset-normalizer==2.0.7
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==4.0.2
commonmark==0.9.1
connexion==2.7.0
croniter==0.3.36
cryptography==3.2.1
defusedxml==0.6.0
dill==0.3.3
dnspython==1.16.0
docutils==0.16
email-validator==1.1.2
Flask==1.1.2
Flask-AppBuilder==3.1.1
Flask-Babel==1.0.0
Flask-Caching==1.9.0
Flask-JWT-Extended==3.25.0
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.13
Flask-WTF==0.14.3
funcsigs==1.0.2
graphviz==0.15
gunicorn==19.10.0
h11==0.12.0
httpcore==0.13.7
httpx==0.20.0
idna==2.10
importlib-metadata==1.7.0
importlib-resources==1.5.0
inflection==0.5.1
iso8601==0.1.13
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
json-merge-patch==0.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.23.1
msal==1.16.0
msal-extensions==0.3.0
msrest==0.6.21
msrestazure==0.6.4
natsort==7.1.0
numpy==1.19.4
oauthlib==3.1.1
openapi-spec-validator==0.2.9
packaging==21.2
pandas==1.1.4
pendulum==2.1.2
portalocker==1.7.1
prison==0.1.3
psutil==5.7.3
psycopg2==2.9.1
pycparser==2.20
Pygments==2.7.2
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-daemon==2.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.4
pytzdata==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rich==9.2.0
setproctitle==1.2
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.3.20
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.36.8
swagger-ui-bundle==0.0.8
tabulate==0.8.7
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
thrift==0.13.0
typing-extensions==3.7.4.3
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.11
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.0
